requirements.txt
djangorestframework==3.7.7
django==1.11.8
gunicorn==19.7.1
